refactor: move [stationName] to features
