+++ site/src/features/pages/station/index.ts
@@ -1,0 +1,1 @@
+export { Station } from './components/page'
